(minor) update route pages/sample to be pages/samples
